Outpass Webapp + Api's - Django/requirements.txt
asn1crypto==0.24.0
Babel==2.7.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
Django==2.2.4
django-cors-headers==3.1.0
django-jsignature==0.8
django-phone-field==1.7.1
django-phonenumber-field==3.0.1
django-rest-framework==0.1.0
djangorestframework==3.10.2
djangorestframework-simplejwt==4.3.0
html5lib==1.0.1
idna==2.8
pdfcrowd==4.9.1
pdfkit==0.6.1
phonenumbers==8.10.16
Pillow==6.1.0
psycopg2==2.8.3
pycparser==2.19
PyJWT==1.7.1
PyPDF2==1.26.0
pytz==2019.2
reportlab==3.5.23
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
webencodings==0.5.1
xhtml2pdf==0.2.3
